+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+## Breaking Changes
+
+* All google connectors now require `token` to be either set to `{"file": "<path to json>"}` or `{"json": {...}}`
 ## [0.13.0-rc.12]
 
 ### Fixes